+++ src/featureflagservice/lib/featureflagservice/repo.ex
@@ -1,0 +1,5 @@
+defmodule Featureflagservice.Repo do
+  use Ecto.Repo,
+    otp_app: :featureflagservice,
+    adapter: Ecto.Adapters.Postgres
+end