+++ src/featureflagservice/lib/featureflagservice_web/controllers/page_controller.ex
@@ -1,0 +1,10 @@
+defmodule FeatureflagserviceWeb.PageController do
+  use FeatureflagserviceWeb, :controller
+
+  alias Featureflagservice.FeatureFlags
+
+  def index(conn, _params) do
+    featureflags = FeatureFlags.list_feature_flags()
+    render(conn, "index.html", featureflags: featureflags)
+  end
+end