+++ src/featureflagservice/lib/featureflagservice_web/views/layout_view.ex
@@ -1,0 +1,3 @@
+defmodule FeatureflagserviceWeb.LayoutView do
+  use FeatureflagserviceWeb, :view
+end