+++ src/featureflagservice/lib/featureflagservice_web/views/page_view.ex
@@ -1,0 +1,3 @@
+defmodule FeatureflagserviceWeb.PageView do
+  use FeatureflagserviceWeb, :view
+end