+++ src/frontend/main.go
@@ -30,8 +30,15 @@ import (
 	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
 
 	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric"
+	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetricgrpc"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracegrpc"
+	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/global"
 	"go.opentelemetry.io/otel/propagation"
+	controller "go.opentelemetry.io/otel/sdk/metric/controller/basic"
+	processor "go.opentelemetry.io/otel/sdk/metric/processor/basic"
+	"go.opentelemetry.io/otel/sdk/metric/selector/simple"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 )
 
@@ -96,21 +103,53 @@ func InitTracerProvider() *sdktrace.TracerProvider {
 	return tp
 }
 
+func InitMeterProvider() (metric.MeterProvider, error) {
+	ctx := context.Background()
+
+	client := otlpmetricgrpc.NewClient(
+		otlpmetricgrpc.WithInsecure())
+
+	metricExporter, err := otlpmetric.New(ctx, client)
+	if err != nil {
+		log.Errorf("failed to create the metric collector exporter, %v", err)
+		return nil, err
+	}
+
+	controller := controller.New(
+		processor.NewFactory(
+			simple.NewWithHistogramDistribution(),
+			metricExporter,
+		),
+		controller.WithExporter(metricExporter),
+		controller.WithCollectPeriod(2*time.Second),
+	)
+
+	err = controller.Start(ctx)
+	if err != nil {
+		log.Errorf("failed to start metric controller, %v", err)
+	}
+	global.SetMeterProvider(controller)
+
+	return controller, nil
+}
+
 func main() {
+	// Initialize OpenTelemetry Tracing
 	tp := InitTracerProvider()
 	defer func() {
 		if err := tp.Shutdown(context.Background()); err != nil {
 			log.Printf("Error shutting down tracer provider: %v", err)
 		}
 	}()
+	log.Printf("initialize trace provider done")
+
+	_, err := InitMeterProvider()
+	if err != nil {
+		log.Errorf("Error init meter provider: %v", err)
+	}
+	log.Printf("initialize meter provider done")
 
 	ctx := context.Background()
-
-	srvPort := port
-	if os.Getenv("PORT") != "" {
-		srvPort = os.Getenv("PORT")
-	}
-	addr := os.Getenv("LISTEN_ADDR")
 	svc := new(frontendServer)
 	mustMapEnv(&svc.productCatalogSvcAddr, "PRODUCT_CATALOG_SERVICE_ADDR")
 	mustMapEnv(&svc.currencySvcAddr, "CURRENCY_SERVICE_ADDR")
@@ -129,6 +168,7 @@ func main() {
 	mustConnGRPC(ctx, &svc.adSvcConn, svc.adSvcAddr)
 
 	r := mux.NewRouter()
+	// Add OpenTelemetry instrumentation to incoming HTTP requests controlled by the gorilla/mux Router.
 	r.Use(otelmux.Middleware("server"))
 	r.HandleFunc("/", instrumentHandler(svc.homeHandler)).Methods(http.MethodGet, http.MethodHead)
 	r.HandleFunc("/product/{id}", instrumentHandler(svc.productHandler)).Methods(http.MethodGet, http.MethodHead)
@@ -146,8 +186,11 @@ func main() {
 	handler = &logHandler{log: log, next: handler} // add logging
 	handler = ensureSessionID(handler)             // add session ID
 
-	log.Infof("starting server on " + addr + ":" + srvPort)
-	log.Fatal(http.ListenAndServe(addr+":"+srvPort, handler))
+	var addr string
+	mustMapEnv(&addr, "FRONTEND_ADDR")
+
+	log.Infof("starting server on " + addr)
+	log.Fatal(http.ListenAndServe(addr, handler))
 }
 
 func mustMapEnv(target *string, envKey string) {
@@ -159,6 +202,7 @@ func mustMapEnv(target *string, envKey string) {
 }
 
 func mustConnGRPC(ctx context.Context, conn **grpc.ClientConn, addr string) {
+	// Add OpenTelemetry instrumentation to outgoing gRPC requests
 	var err error
 	ctx, cancel := context.WithTimeout(ctx, time.Second*3)
 	defer cancel()